Create What I'm thinking
I know you haven't read the assignment as closely as I did, but this is just a quick glance as to what it is I am thinking.  And just to let you know I already made changes to what I had on the function list.  In my opinion we won't really know what is easier until we actually start coding though.